add logic to remove authUser from dummy event and assign to an attendee
