+++ src/utils/helpers/calendarHelpers.ts
@@ -2,11 +2,54 @@ import dayjs from 'dayjs'
 import utc from 'dayjs/plugin/utc'
 import timezone from 'dayjs/plugin/timezone'
 import { DateFieldsInterface, MeetingModalInfo } from 'interfaces'
+import { staticEmails } from 'utils/data/calendarConstants'
 
 dayjs.extend(utc)
 dayjs.extend(timezone)
 
 export const blankDayJs = () => dayjs()
+
+export const buildSandboxEvent = (eventInfo, eventId?) => {
+  const { description, end, start, summary } = eventInfo
+  const updatedAttendees = eventInfo.attendees.map(member => {
+    return { ...member, responseStatus: 'needsAction' }
+  })
+
+  if (!staticEmails.includes(updatedAttendees[0].email)) {
+    updatedAttendees.shift()
+  }
+
+  if (!updatedAttendees.find(attendee => attendee.comment === 'organizer')) {
+    updatedAttendees[0] = { ...updatedAttendees[0], comment: 'organizer' }
+  }
+
+  const startTime = changeDateTimeZone(start.dateTime, start.timezone)
+  const endTime = changeDateTimeZone(end.dateTime, end.timezone)
+
+  return {
+    attendees: updatedAttendees,
+    description,
+    creator: {
+      email: 'erick.manrique@bootcampr.io',
+    },
+    end: endTime,
+    eventId: eventId || generateHexadecimal(),
+    googleDateFields: {
+      endTime: end.dateTime,
+      startTime: start.dateTime,
+    },
+    hangoutLink: 'https://meet.google.com',
+    start: startTime,
+    timeZone: start.timeZone,
+    title: summary,
+  }
+}
+
+export const changeDateTimeZone = (time, timeZone) => {
+  const formattedTime = dayjs(time).format('YYYY-MM-DDTHH:mm')
+  const timeZonedTime = dayjs.tz(formattedTime, timeZone).format()
+  return timeZonedTime
+}
 
 export const checkIfAllMembersInvited = (
   attendees,
@@ -116,12 +159,6 @@ const generateInitialTime = type => {
   }
 
   return dayjs(finalStr).toISOString()
-}
-
-export const changeDateTimeZone = (time, timeZone) => {
-  const formattedTime = dayjs(time).format('YYYY-MM-DDTHH:mm')
-  const timeZonedTime = dayjs.tz(formattedTime, timeZone).format()
-  return timeZonedTime
 }
 
 export const generateHexadecimal = () => {
